bloom/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Bloom backend"
requires-python = ">=3.12"
dependencies = [
    "alembic==1.14.0",
    "annotated-types==0.7.0",
    "anyio==4.6.2.post1",
    "asyncpg==0.30.0",
    "click==8.1.7",
    "dnspython==2.7.0",
    "ecdsa==0.19.0",
    "email_validator==2.2.0",
    "fastapi==0.115.4",
    "greenlet==3.1.1",
    "h11==0.14.0",
    "httptools==0.6.4",
    "idna==3.10",
    "passlib==1.7.4",
    "pyasn1==0.6.1",
    "pydantic==2.9.2",
    "pydantic-settings==2.6.1",
    "pydantic_core==2.23.4",
    "python-dotenv==1.0.1",
    "python-jose==3.3.0",
    "python-multipart==0.0.17",
    "PyYAML==6.0.2",
    "rsa==4.9",
    "six==1.16.0",
    "sniffio==1.3.1",
    "SQLAlchemy==2.0.36",
    "starlette==0.41.2",
    "typing_extensions==4.12.2",
    "uvicorn==0.32.0",
    "uvloop==0.21.0",
    "watchfiles==0.24.0",
    "websockets==14.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api"]

[tool.hatch.metadata]
allow-direct-references = true

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio",
    "black",
    "isort",
    "flake8",
    "mypy",
]

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.isort]
profile = "black"
multi-line-output = 3 